requirements.txt
torch
torchvision
torchaudio
tensorflow
scikit-learn
pillow
matplotlib
tqdm
python-dotenv
